requirements/local.txt
-r base.txt

django-debug-toolbar==1.4
flake8==2.5.4